Update free slot calculation in create_reservation
Resolves a bug in the way free slots are calculated when creating a
reservation. Previously, all allocated and active reservations between
start and end dates of the new proposed reservation were counted and
compared to the slot count of the flavor. But that was too simplistic and
doesn't account for scenarios where multiple shorter reservations exist
within the timeframe.

The flavor_free_slots function performs the correct calculation so use
that instead of recreating the same logic in create_reservation.

Change-Id: I4b067237049a6396b7ab3d159e1fdd58737012e8